refactor(daemon): carve lifecycle sub-package
Move three self-contained files out of internal/daemon into a new
internal/daemon/lifecycle sub-package:

  - scheduler.go       — cron-style Scheduler (gocron wrapper)
  - worker_group.go    — WorkerGroup (background goroutine supervision)
  - stop_context.go    — stop-aware context helper

Plus the matching scheduler_test.go.

The package has no daemon-state dependencies (0 *Daemon methods, 0
d.X refs). stopAwareContext was a method on *Daemon that only
read d.stopChan; convert it to an exported free function
lifecycle.StopAwareContext(parent, stop).

To avoid rewriting every call site, internal/daemon/lifecycle_aliases.go
declares type aliases:

  type Scheduler   = lifecycle.Scheduler
  type WorkerGroup = lifecycle.WorkerGroup

The aliases resolve to the same underlying types, so existing field
declarations (scheduler *Scheduler, workers WorkerGroup) compile
unchanged. NewScheduler() is a thin wrapper around
lifecycle.NewScheduler() so the test suite and NewDaemonWithConfigFile
don't need to be updated.

Two call sites updated to use the new package directly:
  - daemon.go: d.stopAwareContext(ctx) -> lifecycle.StopAwareContext(ctx, d.stopChan)
  - daemon_loop.go: same

This is T20-leaf carve 1 of 6.

Verification: all 45 packages pass; golangci-lint clean; no
fmt.Errorf introduced. internal/daemon/ now has 27 non-test
production files (target: ≤15 after all 6 carves).
