+++ internal/daemon/lifecycle/scheduler_test.go
@@ -1,4 +1,4 @@
-package daemon
+package lifecycle
 
 import (
 	"context"
