+++ internal/daemon/lifecycle/stop_context.go
@@ -1,23 +1,21 @@
-package daemon
+package lifecycle
 
-import (
-	"context"
-)
+import "context"
 
-// stopAwareContext returns a context that is canceled when either the parent
-// context is done or the daemon stop channel is closed.
+// StopAwareContext returns a context that is canceled when either the parent
+// context is done or the supplied stop channel is closed.
 //
-// This preserves the historical behavior where closing stopChan unblocks
-// in-flight work even when the caller passes context.Background().
+// This preserves the historical behavior where closing the stop channel
+// unblocks in-flight work even when the caller passes context.Background().
 //
 // Callers MUST call the returned cancel func when the derived context is no
 // longer needed; otherwise the internal stop-listener goroutine may live for
 // the lifetime of the parent context.
-func (d *Daemon) stopAwareContext(parent context.Context) (context.Context, context.CancelFunc) {
+func StopAwareContext(parent context.Context, stop <-chan struct{}) (context.Context, context.CancelFunc) {
 	if parent == nil {
 		parent = context.Background()
 	}
-	if d == nil || d.stopChan == nil {
+	if stop == nil {
 		ctx, cancel := context.WithCancel(parent)
 		return ctx, cancel
 	}
@@ -25,7 +23,7 @@ func (d *Daemon) stopAwareContext(parent context.Context) (context.Context, cont
 	ctx, cancel := context.WithCancel(parent)
 	go func() {
 		select {
-		case <-d.stopChan:
+		case <-stop:
 			cancel()
 		case <-ctx.Done():
 			// parent canceled; nothing else to do
